@@ -13,11 +13,11 @@
     "php": ">=7.4",
     "ext-json": "*",
     "doctrine/annotations": "^1.13",
-    "symfony/config": "^5.3",
-    "symfony/dependency-injection": "^5.3",
-    "symfony/event-dispatcher": "^5.3",
-    "symfony/http-foundation": "^5.3",
-    "symfony/http-kernel": "^5.3",
+    "symfony/config": "^6.0",
+    "symfony/event-dispatcher": "^6.0",
+    "symfony/dependency-injection": "^6.0",
+    "symfony/http-foundation": "^6.0",
+    "symfony/http-kernel": "^6.0",
     "oas-php/utils": "dev-master",
     "oas-php/resolver": "dev-master",
     "oas-php/schema": "dev-master",
